requirements.txt
PyYAML==6.0.1
pycodestyle==2.11.1
pyflakes==3.2.0
paramiko==3.4.0
pylint==3.0.3
requests==2.31.0
